Update SDK dependencies (#100)
* Update dependencies

* Fix tests

* Fix field assignment

* Update field assignment

* Remove unused functions

---------

+++ stackit/internal/services/postgresflex/user/resource_test.go
@@ -51,7 +51,7 @@ func TestMapFieldsCreate(t *testing.T) {
 					Username: utils.Ptr("username"),
 					Password: utils.Ptr("password"),
 					Host:     utils.Ptr("host"),
-					Port:     utils.Ptr(int32(1234)),
+					Port:     utils.Ptr(int64(1234)),
 				},
 			},
 			Model{
@@ -80,7 +80,7 @@ func TestMapFieldsCreate(t *testing.T) {
 					Username: nil,
 					Password: utils.Ptr(""),
 					Host:     nil,
-					Port:     utils.Ptr(int32(2123456789)),
+					Port:     utils.Ptr(int64(2123456789)),
 				},
 			},
 			Model{
@@ -185,7 +185,7 @@ func TestMapFields(t *testing.T) {
 					},
 					Username: utils.Ptr("username"),
 					Host:     utils.Ptr("host"),
-					Port:     utils.Ptr(int32(1234)),
+					Port:     utils.Ptr(int64(1234)),
 				},
 			},
 			Model{
@@ -212,7 +212,7 @@ func TestMapFields(t *testing.T) {
 					Roles:    &[]string{},
 					Username: nil,
 					Host:     nil,
-					Port:     utils.Ptr(int32(2123456789)),
+					Port:     utils.Ptr(int64(2123456789)),
 				},
 			},
 			Model{
